requirements.txt
django >3.2,<3.3
python-decouple
pillow
djangorestframework
markdown       
django-filter  
drf-yasg[validation]
django-coreapi
djangorestframework-simplejwt
django-mail-templated
djoser
black
flake8